requirements.txt
asgiref==3.5.0
Django==4.0.3
djangorestframework==3.13.1
pkg_resources==0.0.0
pytz==2022.1
sqlparse==0.4.2
